Finished up table & progress ; Updated README ; Fixed fallback to list with -h or --help flags

+++ table_col.go
@@ -12,8 +12,8 @@ var (
 	rxPrefixSplit = regexp.MustCompile(`^(\s+)(.+)$`)
 	rxRightTrim   = regexp.MustCompile(`\s+$`)
 
-	TableColWrapper = func(limit uint) *Wrapper {
-		wrapper := NewWrapper(limit)
+	TableColWrapper = func(limit int) *Wrapper {
+		wrapper := NewWrapper(uint(limit))
 		wrapper.WhitespaceMode = WRAP_WHITESPACE_CONTRACT
 		wrapper.TrimMode = WRAP_TRIM_RIGHT
 		wrapper.BreakWords = true
@@ -29,14 +29,14 @@ func NewTableCol(row *TableRow, content string) *TableCol {
 
 // Render returns wrapped content, max content line width and content line count.
 // See `LineCount()`, `Width()` and `Content()` for more informations.
-func (this *TableCol) Render(maxWidth uint) (content string, width, lineCount uint) {
+func (this *TableCol) Render(maxWidth int) (content string, width, lineCount int) {
 	content = this.Content(maxWidth)
 	lines := strings.Split(content, "\n")
 	rendered := make([]string, len(lines))
 	for idx, line := range lines {
 		lineLen := StringLength(line)
-		if uint(lineLen) > width {
-			width = uint(lineLen)
+		if lineLen > width {
+			width = lineLen
 		}
 		if maxWidth > 0 {
 			if diff := int(maxWidth) - lineLen; diff > 0 {
@@ -64,18 +64,18 @@ func (this *TableCol) Render(maxWidth uint) (content string, width, lineCount ui
 //  maxWidth  = 3
 //  content = "foo bar baz"
 //  line count = 3
-func (this *TableCol) LineCount(maxWidth ...uint) uint {
+func (this *TableCol) LineCount(maxWidth ...int) int {
 	if len(maxWidth) == 0 || maxWidth[0] == 0 {
 		return this.lineCount
 	}
-	return uint(strings.Count(this.Content(maxWidth[0]), "\n") + 1)
+	return strings.Count(this.Content(maxWidth[0]), "\n") + 1
 }
 
 // Content returns the string content of the column. If maxWidth is not given
 // or 0 then the original content is returned. Otherwise the returned content
 // is wrapped to the maxWidth limitation. Words will be split if they exceed
 // the maxWidth value.
-func (this *TableCol) Content(maxWidth ...uint) string {
+func (this *TableCol) Content(maxWidth ...int) string {
 	rendered := this.renderedContent()
 	_dbgTableCol("\033m-- RENDERED:\n%s\n\033[0m--\n", rendered)
 	if len(maxWidth) > 0 && maxWidth[0] > 0 {
@@ -85,7 +85,7 @@ func (this *TableCol) Content(maxWidth ...uint) string {
 }
 
 // ContentPrefixed
-func (this *TableCol) ContentPrefixed(prefix string, maxWidth ...uint) string {
+func (this *TableCol) ContentPrefixed(prefix string, maxWidth ...int) string {
 	lines := strings.Split(this.Content(maxWidth...), "\n")
 	for i, line := range lines {
 		lines[i] = prefix + line
@@ -93,7 +93,7 @@ func (this *TableCol) ContentPrefixed(prefix string, maxWidth ...uint) string {
 	return strings.Join(lines, "\n")
 }
 
-func (this *TableCol) Width(maxWidth ...uint) uint {
+func (this *TableCol) Width(maxWidth ...int) int {
 	return this.maxLineWidth(maxWidth...)
 }
 
@@ -101,7 +101,7 @@ func (this *TableCol) SetContent(content string) *TableCol {
 	content = rxRightTrim.ReplaceAllString(content, "")
 	this.content = &content
 	lines := strings.Split(content, "\n")
-	this.lineCount = uint(len(lines))
+	this.lineCount = len(lines)
 	return this
 }
 
@@ -110,17 +110,17 @@ func (this *TableCol) SetRenderer(renderer func(string) string) *TableCol {
 	return this
 }
 
-func (this *TableCol) maxLineWidth(maxWidth ...uint) (width uint) {
+func (this *TableCol) maxLineWidth(maxWidth ...int) (width int) {
 	rendered := this.renderedLines(maxWidth...)
 	for _, line := range rendered {
-		if l := uint(StringLength(line)); l > width {
+		if l := StringLength(line); l > width {
 			width = l
 		}
 	}
 	return
 }
 
-func (this *TableCol) renderedLines(maxWidth ...uint) []string {
+func (this *TableCol) renderedLines(maxWidth ...int) []string {
 	return strings.Split(this.Content(maxWidth...), "\n")
 }
