+++ table_row.go
@@ -6,61 +6,61 @@ import (
 )
 
 func NewTableRow(cols []string) *TableRow {
-	l := uint(len(cols))
+	l := len(cols)
 	this := &TableRow{
 		ColAmount: l,
 		Cols:      make([]*TableCol, l),
 	}
 	for i, col := range cols {
-		this.SetCol(uint(i), NewTableCol(this, col))
+		this.SetCol(i, NewTableCol(this, col))
 	}
 	return this
 }
 
-func (this *TableRow) Render(totalWidth uint) (rendered []string, maxLineCount uint) {
+func (this *TableRow) Render(totalWidth int) (rendered []string, maxLineCount int) {
 	widths := this.CalculateWidths(totalWidth)
 	return this.RenderWithWidths(widths)
 }
 
-func (this *TableRow) Width(maxWidth ...uint) uint {
-	width := uint(0)
+func (this *TableRow) Width(maxWidth ...int) int {
+	width := 0
 	for _, col := range this.Cols {
 		width += col.Width(maxWidth...)
 	}
 	return width
 }
 
-func (this *TableRow) CalculateWidths(totalWidth uint) (colSize []uint) {
-	colSize = make([]uint, this.ColAmount)
+func (this *TableRow) CalculateWidths(totalWidth int) (colSize []int) {
+	colSize = make([]int, this.ColAmount)
 	factor := float64(1.0)
 	actualWidth := this.Width()
 	if totalWidth > 0 && totalWidth != actualWidth {
 		factor = float64(totalWidth) / float64(actualWidth)
 	}
 	//fmt.Printf("\nTOTAL WIDTH: %d -> %d, FACTOR: %.2f\n", this.TotalWidth, totalWidth, factor)
-	usedWidth := uint(0)
+	usedWidth := 0
 	lastCol := int(this.ColAmount - 1)
 	for idx, col := range this.Cols {
-		width := uint(factor * float64(col.Width()))
+		width := factor * float64(col.Width())
 		//widthPre := width
 		if totalWidth == 0 {
 			width = 0
 		} else {
 			if idx == lastCol { // last col get's it all
 				//fmt.Printf("  @@ LAST COL: TOTAL=%d, USED=%d, WIDTH=%d, FACTOR=%.3f\n", totalWidth, usedWidth, width, factor)
-				width += totalWidth - (usedWidth + width)
+				width += float64(totalWidth - (usedWidth + int(width)))
 			} else if width == 0 { //
 				width = 1
 			} else { // collected used width to spend all in last col
-				usedWidth += width
+				usedWidth += int(width)
 			}
 		}
-		colSize[idx] = width
+		colSize[idx] = int(width)
 	}
 	return
 }
 
-func (this *TableRow) RenderWithWidths(colWidths []uint) (rendered []string, maxLineCount uint) {
+func (this *TableRow) RenderWithWidths(colWidths []int) (rendered []string, maxLineCount int) {
 	rendered = make([]string, this.ColAmount)
 	//fmt.Printf("\nTOTAL WIDTH: %d -> %d, FACTOR: %.2f\n", this.TotalWidth, totalWidth, factor)
 	//fmt.Printf("\nRENDER ROW WITH COL WIDTHS %v\n", colWidths)
@@ -87,7 +87,7 @@ func (this *TableRow) SetTable(table *Table) *TableRow {
 	return this
 }
 
-func (this *TableRow) SetCol(idx uint, col *TableCol) error {
+func (this *TableRow) SetCol(idx int, col *TableCol) error {
 	if idx >= this.ColAmount {
 		return fmt.Errorf("Column index %d is beyond colum size %d", idx, this.ColAmount)
 	}
